uploading and downloading file

+++ controller/admin.js
@@ -1,4 +1,5 @@
 const Product = require("../models/product");
+const fileHelper = require("../util/file");
 
 const { validationResult } = require("express-validator/check");
 
@@ -9,7 +10,7 @@ exports.getAddProduct = (req, res, next) => {
     editing: false,
     hasError: false,
     errorMessage: null,
-    validationErrors: []
+    validationErrors: [],
   }); //for pug
 };
 
@@ -27,9 +28,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   let title = req.body.title;
+  let image = req.file;
   let description = req.body.description;
   let price = req.body.price;
-  let imageUrl = req.body.imageUrl;
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      title: "Add Product",
+      path: "/admin/add-product",
+      editing: false,
+      errorMessage: "The file format is invalid.",
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description,
+      },
+      validationErrors: [],
+    });
+  }
+  let imageUrl = image.path;
   // req.user.createProduct({
   //   title: req.body.title,
   //   description: req.body.description,
@@ -46,21 +63,20 @@ exports.postAddProduct = (req, res, next) => {
   // );
 
   const errors = validationResult(req);
-  if(!errors.isEmpty()){
+  if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
       title: "Add Product",
       path: "/admin/add-product",
       editing: false,
       errorMessage: errors.array()[0].msg,
       hasError: true,
-      product:{
+      product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
-        description: description
+        description: description,
       },
-      validationErrors: errors.array()
-    }); 
+      validationErrors: errors.array(),
+    });
   }
   const product = new Product({
     title: title,
@@ -72,7 +88,12 @@ exports.postAddProduct = (req, res, next) => {
   product
     .save()
     .then(() => res.redirect("/admin/products"))
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+      //return res.redirect('/500')
+    });
 };
 
 // exports.getEditProduct = (req, res, next) => {
@@ -96,7 +117,7 @@ exports.postAddProduct = (req, res, next) => {
 exports.getEditProduct = (req, res, next) => {
   var editMode = req.query.edit;
   const productId = req.params.productId;
- 
+
   Product.findById(productId)
     .then((product) => {
       if (!product) {
@@ -108,10 +129,15 @@ exports.getEditProduct = (req, res, next) => {
         editing: editMode,
         product: product,
         validationErrors: [],
-        errorMessage: ''
+        errorMessage: "",
       });
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+      //return res.redirect('/500')
+    });
 };
 
 // exports.postEditProduct = (req, res, next) => {
@@ -148,6 +174,8 @@ exports.getEditProduct = (req, res, next) => {
 
 exports.postEditProduct = (req, res, next) => {
   const productId = req.body.productId;
+
+  let image = req.file;
   // const product = new Product(
   //   req.body.title,
   //   req.body.price,
@@ -156,22 +184,21 @@ exports.postEditProduct = (req, res, next) => {
   //   req.body.productId,
   //   req.user._id
   // );
-  const errors = validationResult(req)
-  if(!errors.isEmpty()){
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
       title: "Edit Product",
       path: `/admin/edit-product`,
       editing: true,
-      product:{
+      product: {
         _id: productId,
         title: req.body.title,
         description: req.body.description,
-        imageUrl: req.body.imageUrl,
         price: req.body.price,
       },
       validationErrors: errors.array(),
-      errorMessage: errors.array()[0].msg
-    }); 
+      errorMessage: errors.array()[0].msg,
+    });
   }
 
   Product.findOne({ _id: productId, userId: req.user._id })
@@ -180,12 +207,20 @@ exports.postEditProduct = (req, res, next) => {
         product.title = req.body.title;
         product.price = req.body.price;
         product.description = req.body.description;
-        product.imageUrl = req.body.imageUrl;
+        if (image) {
+          fileHelper.deleteProduct(product.imageUrl);
+          product.imageUrl = image.path;
+        }
         product.save();
       }
     })
     .then(() => res.redirect("/admin/products"))
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+      //return res.redirect('/500')
+    });
 };
 
 // exports.deleteProduct = (req, res, next) => {
@@ -205,9 +240,21 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.deleteProduct = (req, res, next) => {
   const productId = req.body.productId;
-  Product.deleteOne({ _id: productId, userId: req.user._id })
+  Product.findById(productId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("No product found"));
+      }
+      fileHelper.deleteProduct(product.imageUrl);
+      return Product.deleteOne({ _id: productId, userId: req.user._id });
+    })
     .then(() => res.redirect("/admin/products"))
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+      //return res.redirect('/500')
+    });
 };
 
 // exports.getProducts = (req, res, next) => {
@@ -253,5 +300,10 @@ exports.getProducts = (req, res, next) => {
         path: "/admin/products",
       });
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+      //return res.redirect('/500')
+    });
 };
